+++ Changelog.md
@@ -2,6 +2,12 @@
 
 Note: This is in reverse chronological order, so newer entries are added to the top.
 
+## Swift Next
+
+- Multiple trailing closure clause syntax has been introduced
+
+  The `FunctionCallExprSyntax.trailingClosure` is now `Syntax?` that can be `ClosureExprSyntax`, `MultipleTrailingClosureClauseSyntax` or `nil`.
+
 ## Swift 5.2
 
 - `ReversedSyntaxChildren` has been removed